src/requirements.txt
Flask
eventlet
python-socketio